Implementation of random forest using data from EtherscanAPI. Added README.md and api_key.
