+++ anomaly/karan_ponzi/api/api_key.json
@@ -1,0 +1,3 @@
+{
+  "key": "JNW16Q1J1Q4SUD9PVB84QIPTBYKM22AG3U"
+}
