+++ anomaly/karan_ponzi/requirements.txt
@@ -1,0 +1,45 @@
+Click==7.0
+Flask==1.1.1
+Flask-RESTful==0.3.7
+Jinja2==2.10.3
+MarkupSafe==1.1.1
+SQLAlchemy==1.3.11
+Werkzeug==0.16.0
+altgraph==0.16.1
+aniso8601==8.0.0
+bdist-mpkg==
+certifi==2019.11.28
+cffi==1.13.2
+chardet==3.0.4
+config==0.4.2
+cryptography==2.8
+cycler==0.10.0
+easydict==1.9
+etherscan==0.3.0
+idna==2.8
+itsdangerous==1.1.0
+joblib==0.14.0
+kiwisolver==1.1.0
+macholib==1.11
+matplotlib==3.1.2
+modulegraph==0.17
+nose==1.3.7
+numpy==1.17.4
+pandas==0.25.3
+pip==19.3.1
+psycopg2==2.8.4
+py2app==0.19
+pyOpenSSL==19.1.0
+pycparser==2.19
+pyparsing==2.4.5
+python-dateutil==2.8.1
+pytz==2019.3
+requests==2.22.0
+requests-cache==0.5.2
+scikit-learn==0.22
+scipy==1.3.3
+setuptools==42.0.2
+six==1.13.0
+sklearn==0.0
+tornado==6.0.3
+urllib3==1.25.7
